removed query mapping file
