+++ salesforce/salesforce.test.js
@@ -4,16 +4,25 @@ const {
   getStudents,
 } = require('.');
 
+const SFDC_OPPTY_RECORD_ID = '012j0000000qVAP';
+
 const conn = new jsforce.Connection({ loginUrl: process.env.SFDC_LOGIN_URL });
-const { QUERY_SELECT, QUERY_WHERE } = require('./query-mapping');
+
+const queryWhere = (courseStart, courseType) => `RecordTypeId = '${SFDC_OPPTY_RECORD_ID}'
+AND Course_Product__c = 'Web Development'
+AND Course_Start_Date_Actual__c = ${courseStart}
+AND Course_Type__c LIKE '%${courseType}%'`;
 
 describe('getStudents', () => {
   test('Should return the expected length of array', async () => {
     await conn.login(process.env.SFDC_USERNAME, process.env.SFDC_PASSWORD,
       (err, userInfo) => userInfo);
     const studentCount = await conn.sobject('Opportunity')
-      .select(QUERY_SELECT)
-      .where(QUERY_WHERE)
+      .select(`Id, Student__c, Student__r.Name,
+      Student__r.Email, Student__r.Secondary_Email__c, Campus_Formatted__c,
+      Student__r.Github_Username__c, Course_Start_Date_Actual__c, Product_Code__c,
+      StageName, Separation_Status__c, Separation_Type__c`)
+      .where(queryWhere)
       .orderby('CreatedDate', 'DESC')
       .execute((err, res) => res);
     const students = await getStudents();
